@@ -62,14 +62,13 @@
     "@types/node": "24.9.1",
     "@types/node-forge": "1.3.14",
     "@types/sshpk": "1.17.4",
-    "@typescript-eslint/eslint-plugin": "7.18.0",
-    "@typescript-eslint/parser": "7.18.0",
+    "@typescript-eslint/eslint-plugin": "8.46.2",
+    "@typescript-eslint/parser": "8.46.2",
     "aws-cdk-lib": "2.220.0",
     "aws-cloudformation-custom-resource": "5.0.0",
     "constructs": "10.4.2",
-    "eslint": "8.57.1",
-    "eslint-config-prettier": "9.1.2",
-    "eslint-plugin-deprecation": "3.0.0",
+    "eslint": "9.38.0",
+    "eslint-config-prettier": "10.1.8",
     "eslint-plugin-prettier": "5.5.4",
     "jsii": "5.9.10",
     "jsii-pacmak": "1.117.0",
@@ -78,7 +77,8 @@
     "publib": "0.2.1025",
     "regenerator-runtime": "0.14.1",
     "ts-node": "10.9.2",
-    "typescript": "5.9.3"
+    "typescript": "5.9.3",
+    "typescript-eslint": "8.46.2"
   },
   "peerDependencies": {
     "aws-cdk-lib": "^2.219.0",
